Extended size of array to int, fixed tree sort exception

+++ Sorting Test/Drawer.cpp
@@ -4,10 +4,10 @@ Drawer::~Drawer()
 {
 }
 
-void Drawer::DrawArray(char* data, short size)
+void Drawer::DrawArray(char* data, int size)
 {
 	Chart->Series[0]->Points->Clear();
-	for (size_t i = 0; i < size; i++)
+	for (int i = 0; i < size; i++)
 	{
 		Chart->Series[0]->Points->Add(data[i]);
 	}
