+++ Sorting Test/Drawer.h
@@ -7,7 +7,7 @@ public ref class Drawer
 public:
 	Drawer(Chart^ chart):Chart(chart) {}
 	~Drawer();
-	void DrawArray(char* data, short size);
+	void DrawArray(char* data, int size);
 
 private:
 	Chart^ Chart;
